+++ spec/rspec/expectations/config.rb
@@ -1,0 +1,11 @@
+RSpec::configure do |config|
+  config.color = true
+
+  config.include FormattingSupport
+
+  config.expect_with :rspec do |expectations|
+    $default_expectation_syntax = expectations.syntax
+    expectations.syntax = :expect
+    expectations.include_chain_clauses_in_custom_matcher_descriptions = true
+  end
+end