+++ spec/rspec/expectations/fixes/missing_constants.rb
@@ -1,0 +1,5 @@
+ComplexStruct = Struct.new(:args)
+
+def Complex(*args)
+  ComplexStruct.new(args)
+end