+++ spec/rspec/expectations/fixes/not_compatible.rb
@@ -1,0 +1,6 @@
+module RSpec::Support::ShellOut
+end
+
+# Uses forking/threads
+module ::RSpec::Support::InSubProcess
+end