+++ spec/rspec/expectations/spec_helper_opal.rb
@@ -1,0 +1,13 @@
+module FormattingSupport
+  def dedent(string)
+    string.gsub(/^\s+\|/, '').chomp
+  end
+end
+
+
+# Can't use subprocesses, etc.
+module MinitestIntegration
+  def with_minitest_loaded
+    yield
+  end
+end